+++ modules/src/ics04_channel/packet.rs
@@ -8,7 +8,10 @@ use crate::ics04_channel::error::Kind;
 use crate::ics24_host::identifier::{ChannelId, PortId};
 use crate::Height;
 
-use super::handler::send_packet::SendPacketResult;
+use super::handler::{
+    acknowledgement::AckPacketResult, recv_packet::RecvPacketResult, send_packet::SendPacketResult,
+    write_acknowledgement::WriteAckPacketResult,
+};
 
 /// Enumeration of proof carrying ICS4 message, helper for relayer.
 #[derive(Clone, Debug, PartialEq, Eq)]
@@ -23,6 +26,14 @@ pub enum PacketMsgType {
 #[derive(Clone, Debug)]
 pub enum PacketResult {
     Send(SendPacketResult),
+    Recv(RecvPacketResult),
+    WriteAck(WriteAckPacketResult),
+    Ack(AckPacketResult),
+}
+
+#[derive(Clone, Debug)]
+pub enum Receipt {
+    Ok,
 }
 
 impl std::fmt::Display for PacketMsgType {